Xiangrui address remaining comments (#6)
